move cache expiry related functions to AQRT
